add feature to department

+++ src/modules/departments/services/departments.manager.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Department } from "../entities/department.entity";
 import { Repository } from "typeorm";
@@ -11,7 +11,7 @@ export class DepartmentsManagerService {
     constructor(
         @InjectRepository(Department)
         private readonly departmentRepository: Repository<Department>
-    ) {}
+    ) { }
 
     // create new Department
     async create(payload: CreateDepartmentDto): Promise<Department> {
@@ -22,13 +22,13 @@ export class DepartmentsManagerService {
     // findAll Departments
     async findAll(): Promise<Department[]> {
         return await this.departmentRepository.find({
-            order: { createdAt: 'DESC'}
+            order: { createdAt: 'DESC' }
         });
     }
 
     // findOne Department
     async findOne(id: number): Promise<Department> {
-        const department = await this.departmentRepository.findOne({ where: { id }})
+        const department = await this.departmentRepository.findOne({ where: { id } })
         if (!department) {
             throw new NotFoundException(`Department with id: ${id} not found`);
         }
@@ -43,8 +43,43 @@ export class DepartmentsManagerService {
     }
 
     // remove Department 
-    async remove(id: number): Promise<void> {
-        const department = await this.findOne(id);
+    async remove(id: number) {
+        const department = await this.departmentRepository.findOne({
+            where: { id },
+            relations: ['users']
+        });
+
+        if (!department) throw new NotFoundException(`دپارتمان با آیدی ${id} پیدا نشد`);
+
+        if (department.users?.length > 0) {
+            throw new BadRequestException('این دپارتمان کارمند فعال دارد و قابل حذف نیست');
+        }
+
         await this.departmentRepository.remove(department);
+        return { message: 'حذف با موفقیت انجام شد', success: true };
+    }
+
+    // assign employee
+    async assignUsersToDepartment(departmentId: number, userId: number | string) {
+        const targetUserId = typeof userId === 'string' ? parseInt(userId, 10) : userId;
+
+        if (isNaN(targetUserId)) {
+            throw new BadRequestException('آیدی کاربر معتبر نیست');
+        }
+
+        const department = await this.departmentRepository.findOne({ where: { id: departmentId } });
+        if (!department) throw new NotFoundException('دپارتمان مورد نظر یافت نشد');
+
+        await this.departmentRepository
+            .createQueryBuilder()
+            .update('users') // اسم جدول یوزرها
+            .set({ department: { id: departmentId } })
+            .where("id = :userId", { userId: targetUserId })
+            .execute();
+
+        return {
+            success: true,
+            message: ` کاربر با موفقیت به دپارتمان ${department.name} منتقل شدند`
+        };
     }
 }
